+++ src/@orbitmines/explorer/Ray.ts
@@ -608,6 +608,88 @@ export class Ray
     // JS.String
     toString = (): string => this.as_array().toString();
     as_string = () => this.toString();
+
+  /**
+   * Quick dirty compilation
+   */
+  to_wolfram_language = (): string => {
+    const hyperEdges: string[][] = [];
+    const options: any = {};
+    let label = 0;
+
+    const vertexStyle = (reference: Option<Ray>): string => {
+      switch (reference.force().type()) {
+        case RayType.INITIAL:{
+          return 'Darker@Red'
+        }
+        case RayType.TERMINAL: {
+          return 'Lighter@Red'
+        }
+        case RayType.REFERENCE: {
+          return 'Orange'
+        }
+        case RayType.VERTEX: {
+          return 'Lighter@Blue'
+        }
+      }
+    }
+    const get = (reference: Option<Ray>): Option<string> => {
+      const vertex = reference.match({ Some: (ref) => ref.vertex(), None: () => Option.None });
+      if (vertex.is_some() && (vertex.force() as any)["__label"] !== undefined)
+        return Option.Some((vertex.force() as any)["__label"] as string);
+
+      return Option.None;
+    }
+
+    this.compile<string, string[]>({
+      directionality: {
+        new: () => {
+          const edge: string[] = [];
+          hyperEdges.push(edge);
+          return edge;
+        },
+        push_back: function (directionality: string[], ray: Option<string>): void {
+          if (ray.is_some())
+            directionality.push(ray.force());
+        }
+      },
+      convert: function (reference: Option<Ray>): Option<string> {
+        const existing = get(reference);
+        if (existing.is_some())
+          return existing;
+
+        if (reference.is_none() || reference.force().vertex().is_none())
+          return Option.None;
+
+        label++;
+
+        const vertex = reference.force().vertex();
+
+        const name: string = `"${vertex.force().js().match({
+          Some: (js) => js.toString(),
+          None: () => `?`
+        })} (${label})"`;
+
+        (options['VertexStyle'] ??= {})[name] = vertexStyle(reference);
+
+        Object.defineProperty(vertex.force(), "__label", {value: name});
+
+        return Option.Some(name);
+      },
+      get,
+    });
+
+    return `ResourceFunction["WolframModelPlot"][{${hyperEdges
+      .filter(hyperEdge => hyperEdge.length !== 0)
+      .map(hyperEdge =>
+        `{${hyperEdge.join(',')}}`
+      ).join(',')}},VertexLabels->All,${
+        _.map(options, (mapping, option) =>
+          `${option} -> <|${
+            _.map(mapping, (value, key) => `${key} -> ${value}`)
+              .join(',')}|>`)
+          .join(',')}]`;
+  }
 }
 
 export type MapOptions = ((direction: Option<Ray>) => Option<Ray>)
